Implement the "flip" function.
This necessitates plumbing through a callback function, so that SDL can be
called.

So as to not repeat ourselves, extract the code that was used around update()
calls, FPS handling, etc, and have a shared function.

While around, reshuffle some initialisation to be earlier, and make it so
there's not a 1/FPS gap before drawing a frame, but a 1/FPS gap _between_
frames (assuming non-zero time to calculate the next frame etc).